requirements.txt
requests~=2.27.1
pandas~=1.3.5
SQLAlchemy~=1.4.32
python-dotenv~=0.19.2
SQLAlchemy-Utils~=0.38.2